requirements.txt
asgiref==3.3.1
attrs==20.3.0
awsebcli==3.19.2
bcrypt==3.2.0
blessed==1.17.12
boto3==1.16.52
botocore==1.19.52
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
cryptography==3.3.1
Django==3.1.5
django-storages==1.11.1
docker==4.4.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==2.10
importlib-metadata==3.4.0
jmespath==0.10.0
jsonschema==3.2.0
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.4
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.4.0
